Fix for error when getting sgtk alembic nodes.
The following methods are failing :-
hou.nodeType(hou.ropNodeTypeCategory(), tk_node_type).instances()
hou.nodeType(hou.sopNodeTypeCategory(), tk_node_type).instances()

This is because if there are no nodes of that type, the first part of the line returns None, and raises an error as it has no "instances()" method. 

The nodes need to be gathered first, and if there are are any, THEN get their instances.